% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmoss.R
\docType{package}
\name{MRMOSS-package}
\alias{MRMOSS}
\alias{MRMOSS-package}
\title{MR-MOSS package}
\description{
MR-MOSS provides R wrappers around the likelihood-based C++ core for
correlated-outcome Mendelian randomization with summary statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YunlongCao/MRMOSS}
  \item Report bugs at \url{https://github.com/YunlongCao/MRMOSS/issues}
}

}
\author{
\strong{Maintainer}: Yunlong Cao \email{157203291+YunlongCao@users.noreply.github.com}

}
\keyword{internal}
